refactor(gadgets.ts): move testRot function and related tests to gadgets.unit-test.ts
